ninth commit

+++ main_app/models.py
@@ -109,7 +109,7 @@ class Batterymodel(models.Model):
     length = models.IntegerField(blank=False)
     width = models.IntegerField(blank=False)
     height = models.IntegerField(blank=False)
-    weight = models.IntegerField(blank=False)
+    weight = models.FloatField(blank=False)
     BH_type = models.CharField(max_length=10, blank=False)
     cell_layout = models.CharField(max_length=20, blank=False)
     model_pic = models.ImageField(null=True, blank=False)
